refactor: 서비스 전부 리팩토링
- s2 계층으로 리팩토링 // 연결 흐름은 그 전 계층에서 진행

+++ src/main/java/com/sprint/mission/discodeit/service/channel/ChannelServiceImpl.java
@@ -2,13 +2,10 @@ package com.sprint.mission.discodeit.service.channel;
 
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.ChannelType;
-import com.sprint.mission.discodeit.entity.ReadStatus;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
-import com.sprint.mission.discodeit.service.readstatus.ReadStatusService;
-import com.sprint.mission.discodeit.web.controller.channel.dto.PrivateChannelCreateRequestDTO;
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
-import lombok.Locked.Read;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -26,9 +23,7 @@ import org.springframework.stereotype.Service;
     하면됨.
  */
 
-/*
-    todo 채널 소유권자를 추가해야할 것 같긴함.
- */
+
 @RequiredArgsConstructor
 @Service
 public class ChannelServiceImpl implements ChannelService {
@@ -36,26 +31,51 @@ public class ChannelServiceImpl implements ChannelService {
 
     @Override
     public Channel makeChannel(Channel channel) {
-
+        if(Objects.isNull(channel)){
+            throw new IllegalArgumentException("채널이 널입니다"); //todo
+        }
         return channelRepository.saveEntity(channel);
     }
 
     @Override
     public Channel findChannelById(UUID channelId) {
+        if(Objects.isNull(channelId)){
+            throw new IllegalArgumentException("채널이 널입니다");
+        }
+
         return channelRepository.findById(channelId)
-            .orElseThrow(() -> new RuntimeException("해당 ID의 채널이 존재하지 않습니다"));
+            .orElseThrow(() -> new IllegalArgumentException("해당 ID의 채널이 존재하지 않습니다"));
     }
 
     @Override
     public Channel updateChannelName(UUID channelId, String updateName) {
+        if(Objects.isNull(channelId) || Objects.isNull(updateName)){
+            throw new IllegalArgumentException("인자는 널일 수 없음");
+        }
+
+        //채널 조회 검증
         Channel channel = this.findChannelById(channelId);
-        channel.update(updateName);
+
+        /*
+            todo 질문리스트 --------
+            근데 이런 비즈니스적 요구사항 처리는 s2에서 하는게 맞나요?
+            초기 요구사항이 퍼블릭 채널 프라이빗 채널 메서드를 구분하는거였는데
+            프론트와 컨트롤러에서 이를 처리할수도 있다 생각해서요
+         */
+        if(channel.getChannelType().equals(ChannelType.PRIVATE_CHANNEL)){
+            throw new IllegalArgumentException("프라이빗 채널은 수정이 불가");
+        }
+        channel.updateChannelName(updateName);
 
         return channelRepository.saveEntity(channel);
     }
 
     @Override
     public void deleteChannel(UUID channelId) {
+        if(Objects.isNull(channelId)){
+            throw new IllegalArgumentException("인자는 널일 수 없음");
+        }
+
         Channel channel = this.findChannelById(channelId);
         channelRepository.deleteEntity(channel.getId());
     }
